warnings reduced; WHAT THE FUCK THIS 'esko' VAR?!! some sort of assert()?

+++ src/game/miscitem.cpp
@@ -2732,7 +2732,7 @@ truth holyhandgrenade::CalculateHasBe() const
 
 void holyhandgrenade::Be() {
   item::Be();
-  if(3 * (game::GetTick() - PinPulledTick) > Count * 100)
+  if(3 * (game::GetTick() - PinPulledTick) > (ulong)(Count) * 100)
   {
     ++Count;
     festring Msg = "A voice loudly declares: \"";
